+++ outscale/resource_outscale_nic_private_ip.go
@@ -65,7 +65,7 @@ func resourceOutscaleOAPINetworkInterfacePrivateIPCreate(d *schema.ResourceData,
 	}
 
 	if v, ok := d.GetOk("secondary_private_ip_count"); ok {
-		input.SetSecondaryPrivateIpCount(int32(v.(int) - 1))
+		input.SetSecondaryPrivateIpCount(int64(v.(int) - 1))
 	}
 
 	if v, ok := d.GetOk("private_ips"); ok {
